gesture: propagate taps through gesture areas
When a gesture is started, its starting point will be saved. This
event and all consecutive events will be stored in a queue. Upon
moving the gesture point, the new point is compared to the old point.
If their difference on the gesture's axis exceeds a certain
threshold, a confirmation flag will be set. When ending the gesture, if
this flag was not set during the gesture, the stored event queue will be
propagated to the active client. Finally, the queue and the
confirmation flag will be cleared.

Additionally, only hide the input panel when the gesture is confirmed
and disable the bottom gesture when the input panel is visible.

Fixes #2